+++ cli/__init__.py
@@ -2,3 +2,5 @@
 Interactive CLI to store scripts into
 '''
 __version__ = '1.0.0'
+
+from cli.logging import Logger  # noqa