+++ cli/app.py
@@ -2,29 +2,29 @@
 CLI application supporting plugins to centralize
 scripts and other odds and ends.
 '''
-
-import importlib.util
-from pathlib import Path
-import sys
+import logging
 
 import typer
 
-from cli.config import plugins_path, plugins_registry, state  # noqa
+from cli.config import plugins_registry, state  # noqa
+from cli.logging import Logger
 from cli.subcommands import env
 from cli.subcommands import config as cfg
+from cli.subcommands import plugins
+from cli.utils import clean_help_string, do_import
 
 
-def do_import(module_name: str, module_path: Path):
-    spec = importlib.util.spec_from_file_location(module_name, str(module_path))
-    module = importlib.util.module_from_spec(spec)
-    sys.modules[spec.name] = module
-    spec.loader.exec_module(module)
-    return module
+logging.setLoggerClass(Logger)
+logger = Logger('cli')
 
 
 def add_plugins_hook(app: typer.Typer):
     for module_name, cached_module in plugins_registry.items():
         # Cached module is in the form {"path": "...", "metadata": {...}}
+        logger.info(
+            'Loading plugin %s into context with data %s',
+            module_name, cached_module
+        )
         module = do_import(module_name, cached_module['path'])
         app.add_typer(module.app, **cached_module['metadata'])
     return app
@@ -32,32 +32,6 @@ def add_plugins_hook(app: typer.Typer):
 
 app = typer.Typer(help=__doc__)
 app = add_plugins_hook(app)
-
-
-@app.command()
-def cache_plugins():
-    for module_path in plugins_path.iterdir():
-        if module_path.isdir() or module_path.name.endswith('.py'):
-            if module_path.name == '__pycache__':
-                continue
-            if module_path.name.startswith('.'):
-                continue
-            module_name = module_path.name.split('.')[0]
-            module_name = f'plugins.{module_name}'
-            module = do_import(module_name, module_path)
-            default_metadata = {
-                'help': module.__doc__,
-                'name': module_path.name.split('.')[0],
-            }
-            metadata = getattr(module, 'metadata', default_metadata)
-            if 'help' not in metadata:
-                metadata['help'] = module.__doc__
-            if 'name' not in metadata:
-                metadata['name'] = module_path.name.split('.')[0]
-            plugins_registry[module_name] = {
-                'path': str(module_path),
-                'metadata': metadata,
-            }
 
 
 @app.callback()
@@ -75,8 +49,9 @@ def callback(
     return
 
 
-app.add_typer(env.app, name='env', help=env.__doc__)
-app.add_typer(cfg.app, name='config', help=cfg.__doc__)
+app.add_typer(env.app, name='env', help=clean_help_string(env.__doc__))
+app.add_typer(cfg.app, name='config', help=clean_help_string(cfg.__doc__))
+app.add_typer(plugins.app, name='plugins', help=clean_help_string(plugins.__doc__))
 
 typer_click_object = typer.main.get_command(app)