+++ cli/config.py
@@ -3,7 +3,23 @@ from pathlib import Path
 from xdgconfig import JsonConfig
 
 
-config = JsonConfig('cli', 'config.json')
+class DefaultConfig:
+    _DEFAULTS = {
+        'app.theme': 'monokai',
+        'app.logger.level': 'info',
+        'app.editor.prefer_visual': False,
+        'app.editor.name': 'vi',
+        'app.editor.flags': [],
+        'app.pager.name': 'less',
+        'app.pager.flags': [],
+    }
+
+
+class Config(DefaultConfig, JsonConfig):
+    ...
+
+
+config = Config('cli', 'config.json')
 env = JsonConfig('cli', 'env.json')
 root_path = Path(__file__).parent.resolve()
 templates_path: Path = (config.app_path / 'templates')
@@ -17,3 +33,6 @@ state = {
     'toml': False,
     'yaml': False,
 }
+RESERVED_COMMANDS = ['config', 'env', 'plugins']
+logs_path: Path = (config.app_path / 'logs')
+logs_path.mkdir(parents=True, exist_ok=True)