+++ CHANGELOG.md
@@ -1,1 +1,12 @@
 # 1.0.0
+
+* Initial version, including :
+  * Extensible plugins system
+    * cache plugins
+    * view plugin logs
+    * edit plugins
+  * Automatic configuration handling
+  * Environment system
+  * Added logging capabilities
+  * Added state management to the CLI (store global flags)
+  * Add default config values